Redo entirely frame interpolation + various fixes
Pretty massive commit.

Sprites have now two nx/ny fields that represents their position on the
next true frame (as opposed to view frame). These are modified by
actions and are used to interpolate the position of the sprites on the
next view frames before the next true frame. On each true frame, nx/ny
are put into x/y, which are the position used by the game logic (meaning
: collisions) and are recalculated by actions.